+++ src/main/java/shop/tryit/web/answer/AnswerAdapter.java
@@ -1,6 +1,7 @@
 package shop.tryit.web.answer;
 
 import shop.tryit.domain.answer.Answer;
+import shop.tryit.domain.answer.dto.AnswerFormDto;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.question.Question;
