+++ src/main/java/shop/tryit/web/question/QuestionController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import shop.tryit.domain.answer.dto.AnswerFormDto;
 import shop.tryit.domain.common.Pages;
 import shop.tryit.domain.question.QuestionSearchCondition;
 import shop.tryit.domain.question.QuestionSearchDto;
@@ -23,7 +24,6 @@ import shop.tryit.domain.question.dto.QuestionCheckPasswordFormDto;
 import shop.tryit.domain.question.dto.QuestionFormDto;
 import shop.tryit.domain.question.dto.QuestionSaveFormDto;
 import shop.tryit.domain.question.service.QnAFacade;
-import shop.tryit.web.answer.AnswerFormDto;
 
 @Slf4j
 @Controller
