+++ src/test/java/shop/tryit/domain/answer/AnswerServiceTests.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.answer.service.AnswerService;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
 import shop.tryit.domain.question.Question;
@@ -49,7 +50,7 @@ class AnswerServiceTests {
 
     @Test
     void 수정_기능_테스트() {
-        //given
+        // given
         Member member = Member.builder().build();
         Question question = Question.of("title", "content", member);
         Answer answer = Answer.of("content2", question, member);
